+++ src/core/observable.ts
@@ -1,7 +1,10 @@
-import { IDerivation, IDerivationState } from "./derivation"
+import { IDerivation, IDerivationState, TraceMode } from "./derivation"
 import { globalState } from "./globalstate"
 import { invariant } from "../utils/utils"
-import { runReactions } from "./reaction"
+import { runReactions, Reaction } from "./reaction"
+import { ComputedValue } from "./computedvalue"
+import { getDependencyTree } from "../api/extras"
+import { IDependencyTree } from "../mobx"
 
 export interface IDepTreeNode {
     name: string
@@ -186,7 +189,12 @@ export function propagateChanged(observable: IObservable) {
     let i = observers.length
     while (i--) {
         const d = observers[i]
-        if (d.dependenciesState === IDerivationState.UP_TO_DATE) d.onBecomeStale()
+        if (d.dependenciesState === IDerivationState.UP_TO_DATE) {
+            if (d.isTracing !== TraceMode.NONE) {
+                logTraceInfo(d, observable)
+            }
+            d.onBecomeStale()
+        }
         d.dependenciesState = IDerivationState.STALE
         d.triggeredBy = observable.name
     }
@@ -226,8 +234,48 @@ export function propagateMaybeChanged(observable: IObservable) {
         const d = observers[i]
         if (d.dependenciesState === IDerivationState.UP_TO_DATE) {
             d.dependenciesState = IDerivationState.POSSIBLY_STALE
+            if (d.isTracing !== TraceMode.NONE) {
+                logTraceInfo(d, observable)
+            }
             d.onBecomeStale()
         }
     }
     // invariantLOS(observable, "maybe end");
+}
+
+function logTraceInfo(derivation: IDerivation, observable: IObservable) {
+    console.log(
+        `[mobx.trace] '${derivation.name}' is invalidated due to a change in: '${observable.name}'`
+    )
+    if (derivation.isTracing === TraceMode.BREAK) {
+        const lines = []
+        printDepTree(getDependencyTree(derivation), lines, 1)
+
+        // prettier-ignore
+        new Function(
+`debugger;
+/*
+Tracing '${derivation.name}'
+
+You are entering this break point because derivation '${derivation.name}' is being traced and '${observable.name}' is now forcing it to update.
+Just follow the stacktrace you should now see in the devtools to see precisely what piece of your code is causing this update
+The stackframe you are looking for is at least ~6-8 stack-frames up.
+
+${derivation instanceof ComputedValue ? derivation.derivation.toString() : ""}
+
+The dependencies for this derivation are:
+
+${lines.join("\n")}
+*/
+    `)()
+    }
+}
+
+function printDepTree(tree: IDependencyTree, lines: string[], depth: number) {
+    if (lines.length >= 1000) {
+        lines.push("(and many more)")
+        return
+    }
+    lines.push(`${new Array(depth).join("\t")}${tree.name}`) // MWE: not the fastest, but the easiest way :)
+    if (tree.dependencies) tree.dependencies.forEach(child => printDepTree(child, lines, depth + 1))
 }